requirements.txt
Flask==3.0.0
mysql-connector-python==8.2.0
qrcode[pil]==7.4.2
python-dotenv==1.0.0
Pillow==10.2.0
Werkzeug==3.0.1
